Refractor `MyApp` to its own folder
