+++ mobile/lib/main.dart
@@ -3,9 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mobile/firebase_options.dart';
-import 'package:mobile/providers/supabase/supabase_client_provider.dart';
 import 'package:mobile/router/app_router.dart';
-import 'package:mobile/ui/theme/app_theme.dart';
+import 'package:mobile/ui/app/pages/my_app.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -16,24 +15,11 @@ Future<void> main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
 
-  runApp(const ProviderScope(child: MyApp()));
-}
+  final appRouter = AppRouter();
 
-final _appRouter = AppRouter();
-
-class MyApp extends ConsumerWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    ref.watch(supabaseClientProvider);
-
-    return MaterialApp.router(
-      routerConfig: _appRouter.config(),
-      theme: AppTheme.light,
-      darkTheme: AppTheme.dark,
-      title: 'Flutter Kenya',
-      debugShowCheckedModeBanner: false,
-    );
-  }
+  runApp(
+    ProviderScope(
+      child: MyApp(appRouter: appRouter),
+    ),
+  );
 }
